feat: add new API routes for clarification and execution
- Implemented /api/clarify route with validation and error handling.
- Added /api/execute route using dependency injection for handler.
- Introduced /api/progress routes for streaming and snapshot retrieval.
- Moved complex prompt logic and related helper functions to a new helpers module.
- Updated ESLint configuration to ignore additional dist directories.
- Added a new task in VSCode for validating the repository with linting, type checking, and testing.
